requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.10.0
deprecation==2.1.0
exceptiongroup==1.2.0
Flask==3.0.3
Flask-Cors==4.0.0
fonttools==4.51.0
gotrue==2.4.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
packaging==24.0
pillow==10.3.0
postgrest==0.16.3
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.1
realtime==1.0.4
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.27
six==1.16.0
sniffio==1.3.1
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.1
supafunc==0.4.5
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.2
